+++ deployments/chef/test/integration/with_custom_systemd_node_instrumentation/test.rb
@@ -1,5 +1,5 @@
 libsplunk_path = '/usr/lib/splunk-instrumentation/libsplunk.so'
-java_tool_options = '-javaagent:/usr/lib/splunk-instrumentation/splunk-otel-javaagent.jar'
+node_options = '-r /usr/lib/splunk-instrumentation/splunk-otel-js/node_modules/@splunk/otel/instrument'
 resource_attributes = 'splunk.zc.method=splunk-otel-auto-instrumentation-\d+\.\d+\.\d+-systemd,deployment.environment=test'
 otlp_endpoint = 'http://0.0.0.0:4317'
 
@@ -7,12 +7,29 @@ describe package('splunk-otel-auto-instrumentation') do
   it { should be_installed }
 end
 
+describe npm('@splunk/otel', path: '/usr/lib/splunk-instrumentation/splunk-otel-js') do
+  it { should be_installed }
+end
+
 describe file('/etc/ld.so.preload') do
   its('content') { should_not match /^#{libsplunk_path}$/ }
 end
 
+describe file('/etc/splunk/zeroconfig/java.conf') do
+  it { should_not exist }
+end
+
+describe file('/etc/splunk/zeroconfig/node.conf') do
+  it { should_not exist }
+end
+
+describe file('/usr/lib/splunk-instrumentation/instrumentation.conf') do
+  it { should_not exist }
+end
+
 describe file('/usr/lib/systemd/system.conf.d/00-splunk-otel-auto-instrumentation.conf') do
-  its('content') { should match /^DefaultEnvironment="JAVA_TOOL_OPTIONS=#{java_tool_options}"$/ }
+  its('content') { should_not match /.*JAVA_TOOL_OPTIONS.*/ }
+  its('content') { should match /^DefaultEnvironment="NODE_OPTIONS=#{node_options}"$/ }
   its('content') { should match /^DefaultEnvironment="OTEL_RESOURCE_ATTRIBUTES=#{resource_attributes}"$/ }
   its('content') { should match /^DefaultEnvironment="OTEL_SERVICE_NAME=test"$/ }
   its('content') { should match /^DefaultEnvironment="SPLUNK_PROFILER_ENABLED=true"$/ }