+++ deployments/chef/test/integration/with_default_preload_node_instrumentation/test.rb
@@ -1,9 +1,13 @@
 libsplunk_path = '/usr/lib/splunk-instrumentation/libsplunk.so'
-java_tool_options = '-javaagent:/usr/lib/splunk-instrumentation/splunk-otel-javaagent.jar'
+node_options = '-r /usr/lib/splunk-instrumentation/splunk-otel-js/node_modules/@splunk/otel/instrument'
 resource_attributes = 'splunk.zc.method=splunk-otel-auto-instrumentation-\d+\.\d+\.\d+'
 otlp_endpoint = 'http://127.0.0.1:4317'
 
 describe package('splunk-otel-auto-instrumentation') do
+  it { should be_installed }
+end
+
+describe npm('@splunk/otel', path: '/usr/lib/splunk-instrumentation/splunk-otel-js') do
   it { should be_installed }
 end
 
@@ -15,8 +19,16 @@ describe file('/usr/lib/systemd/system.conf.d/00-splunk-otel-auto-instrumentatio
   it { should_not exist }
 end
 
+describe file('/usr/lib/splunk-instrumentation/instrumentation.conf') do
+  it { should_not exist }
+end
+
 describe file('/etc/splunk/zeroconfig/java.conf') do
-  its('content') { should match /^JAVA_TOOL_OPTIONS=#{java_tool_options}$/ }
+  it { should_not exist }
+end
+
+describe file('/etc/splunk/zeroconfig/node.conf') do
+  its('content') { should match /^NODE_OPTIONS=#{node_options}$/ }
   its('content') { should match /^OTEL_RESOURCE_ATTRIBUTES=#{resource_attributes}$/ }
   its('content') { should_not match /.*OTEL_SERVICE_NAME.*/ }
   its('content') { should match /^SPLUNK_PROFILER_ENABLED=false$/ }