+++ deployments/chef/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## unreleased
 
+## chef-v0.10.0
+
+- Initial support for [Splunk OpenTelemetry for Node.js](https://github.com/signalfx/splunk-otel-js) Auto
+  Instrumentation on Linux:
+  - The Node.js SDK is installed and activated by default if the `with_auto_instrumentation` option is set to `true`
+    and `npm` is found on the node with the `bash -c 'command -v npm'` shell command.
+  - Set the `with_auto_instrumentation_sdks` option to only `%w(java)` to skip Node.js auto instrumentation.
+  - Use the `auto_instrumentation_npm_path` option to specify a custom path for `npm`.
+  - **Note:** This recipe does not manage the installation/configuration of Node.js, `npm`, or Node.js applications.
+
 ## chef-v0.9.0
 
 - **Deprecations**: The `auto_instrumentation_generate_service_name` and `auto_instrumentation_disable_telemetry`